+++ src/Bootstrap.php
@@ -1,6 +1,10 @@
 <?php
 namespace App;
 
+use App\Admin\PluginFacade;
+
+
+
 if (!defined('ABSPATH')) exit;
 
 use App\Routes\ApiRoutes;
@@ -8,5 +12,6 @@ use App\Routes\ApiRoutes;
 class Bootstrap {
     public static function init() {
         new ApiRoutes();
+        PluginFacade::init();
     }
 }